Fixed warning on windows.
git-svn-id: https://svn.opendtect.org/od/trunk@27017 a985855d-d384-4b45-ae33-31a410998c3c

+++ src/Basic/posinfo2d.cc
@@ -157,7 +157,7 @@ void PosInfo::Line2DData::dump( std::ostream& strm, bool pretty ) const
     else
     {
 	strm << lnm_ << '\n';
-	const float fac = SI().zDomain().userFactor();
+	const int fac = SI().zDomain().userFactor();
 	strm << "Z range " << SI().getZUnitString() << ":\t" << fac*zrg_.start
 	     << '\t' << fac*zrg_.stop << "\t" << fac*zrg_.step;
 	strm << "\n\nTrace number\tX-coord\tY-coord" << std::endl;
